Restructure volcano code in separate dir; now able to launch independent of the DA app
