+++ client/plots/volcano.js
@@ -1,0 +1,1 @@
+export * from './volcano/Volcano.ts'
