+++ client/plots/volcano/Volcano.ts
@@ -5,7 +5,7 @@ import { fillTermWrapper } from '#termsetting'
 import { Menu, sayerror } from '#dom'
 import { RxComponentInner } from '../../types/rx.d'
 import { controlsInit } from '../controls'
-import type { DiffAnalysisInteractions } from './interactions/DiffAnalysisInteractions'
+import type { DiffAnalysisInteractions } from '../diffAnalysis/interactions/DiffAnalysisInteractions'
 import { VolcanoModel } from './model/VolcanoModel'
 import { VolcanoViewModel } from './viewModel/VolcanoViewModel'
 import { VolcanoInteractions } from './interactions/VolcanoInteractions'
