+++ client/plots/volcano/view/VolcanoPlotView.ts
